Remove homepage caching from service worker
The homepage caching does not discriminate between the logged in and logged out state of a user. This causes users that see the AN homepage and subsequently login to be returned to the AN homepage. A cache clear is required in order to see the LU homepage.

To fix this the caching line is removed. The fetch event listener is retained as is explained in the comments.